Extract metadata_t mixin to deduplicate account/item tag handling
The original support for metadata tags on account directives (#1681)
duplicated ~156 lines of tag machinery from item.cc into account.cc:
the case-insensitive comparator, set_tag, has_tag/get_tag for both
string and mask forms, the :tag: / Key: / Key:: parser, and the
expression-engine call_scope_t dispatchers.  The duplication had
already drifted -- account_t::parse_tags lacked the leading-semicolon
skip that fixed #1786 for items, so ;;Key: value mis-parsed on
accounts.

Lift the shared storage and behaviour into a new metadata_t mixin in
src/metadata.{h,cc}.  Both item_t and account_t now publicly inherit
it.  metadata_t::parse_metadata_tags handles the colon syntaxes (with
the #1786 skip applied uniformly); item_t::parse_tags overrides only
to add the bracketed-[date] handling that does not apply to accounts.
The has_tag/get_tag/set_tag virtuals come from the base, and
account_t overrides has_tag/get_tag with three-line walks that fall
back to the parent account when inherit=true (matching the existing
post_t -> xact_t pattern, but walking the account tree).

The expression-binding dispatchers also share one implementation:
metadata_has_tag(metadata_t&, call_scope_t&) and metadata_get_tag(...).
The per-class lookup() entries become one-line trampolines that
resolve the metadata target via find_scope<item_t> or
args.context<account_t>() and forward to the shared dispatcher.

Backwards-compat aliases item_t::string_map and item_t::tag_data_t
preserve every existing reference (Python bindings, output.cc,
print.cc, journal.cc).  put_metadata moves to metadata.cc with its
parameter retyped to metadata_t::string_map (which item_t::string_map
aliases unchanged).

Net effect: account.cc/.h drop from +281 to ~+97 lines vs main,
item.cc shrinks by ~270 lines, and the unified parser fixes the
;;Key: value regression on accounts as a side effect.  All 4301
tests pass.
